LayoutTests/http/tests/websocket/tests/hixie76/handshake-fail-by-no-cr-expected.txt

CONSOLE MESSAGE: Status line does not end with CRLF
Handshake should fail when the first line does not end with CRLF.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS connected is false
PASS origin is undefined.
PASS successfullyParsed is true

TEST COMPLETE
